CHG: Split out utils into typecheck and debug
